+++ TapUIKit-iOS/TapUIKit-iOS/Core/Assets/TapCurrencyWidgetView.xib
@@ -72,9 +72,9 @@
                             <nil key="textColor"/>
                             <nil key="highlightedColor"/>
                         </label>
-                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="item" translatesAutoresizingMaskIntoConstraints="NO" id="heq-RV-34D">
+                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="chevrondown" translatesAutoresizingMaskIntoConstraints="NO" id="heq-RV-34D">
                             <rect key="frame" x="93.333333333333329" y="78" width="18" height="34"/>
-                            <color key="tintColor" red="0.98431372549999996" green="0.0" blue="0.070588235289999995" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                            <color key="tintColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                             <constraints>
                                 <constraint firstAttribute="width" secondItem="heq-RV-34D" secondAttribute="height" multiplier="10:19" id="HJh-uN-sul"/>
                                 <constraint firstAttribute="width" constant="18" id="Ip4-Xm-TKU"/>
@@ -124,7 +124,7 @@
         </view>
     </objects>
     <resources>
-        <image name="item" width="128" height="128"/>
+        <image name="chevrondown" width="18" height="18"/>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
         </systemColor>
