+++ TapUIKit-iOS/TapUIKit-iOS/Core/TapCurrencyWidgetView/TriangleView.swift
@@ -1,0 +1,111 @@
+//
+//  TriangleView.swift
+//  TapUIKit-iOS
+//
+//  Created by MahmoudShaabanAllam on 04/06/2023.
+//  Copyright © 2023 Tap Payments. All rights reserved.
+//
+
+import Foundation
+/// Triangle point to up view
+class UpTriangleView : UIView {
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+    }
+    
+    override func draw(_ rect: CGRect) {
+        guard let context = UIGraphicsGetCurrentContext() else { return }
+        let color = TapThemeManager.colorValue(for: "CurrencyWidget.currencyDropDown.backgroundColor") ?? .white
+        context.beginPath()
+        context.move(to: CGPoint(x: rect.minX, y: rect.maxY))
+        context.addLine(to: CGPoint(x: rect.maxX, y: rect.maxY))
+        context.addLine(to: CGPoint(x: (rect.maxX / 2.0), y: rect.minY))
+        context.closePath()
+        context.setFillColor(color.cgColor)
+        context.fillPath()
+    }
+}
+
+/// Triangle point to down view
+class DownTriangleView : UIView {
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+    }
+    
+    override func draw(_ rect: CGRect) {
+        
+        guard let context = UIGraphicsGetCurrentContext() else { return }
+        let color = TapThemeManager.colorValue(for: "CurrencyWidget.currencyDropDown.backgroundColor") ?? .white
+        context.beginPath()
+        context.move(to: CGPoint(x: rect.minX, y: rect.minY))
+        context.addLine(to: CGPoint(x: rect.maxX, y: rect.minY))
+        context.addLine(to: CGPoint(x: (rect.maxX / 2.0), y: rect.maxY))
+        context.closePath()
+        context.setFillColor(color.cgColor)
+        context.fillPath()
+    }
+}
+
+/// Triangle point to left view
+class LeftTriangleView : UIView {
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+    }
+    
+    override func draw(_ rect: CGRect) {
+        
+        guard let context = UIGraphicsGetCurrentContext() else { return }
+        let color = TapThemeManager.colorValue(for: "CurrencyWidget.currencyDropDown.backgroundColor") ?? .white
+        context.beginPath()
+        context.move(to: CGPoint(x: rect.maxX, y: rect.maxY))
+        context.addLine(to: CGPoint(x: rect.maxX, y: rect.minY))
+        context.addLine(to: CGPoint(x: rect.minX, y: (rect.maxY / 2.0)))
+        context.closePath()
+        context.setFillColor(color.cgColor)
+        context.fillPath()
+    }
+}
+
+/// Triangle point to right view
+class RightTriangleView : UIView {
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+    }
+    
+    override func draw(_ rect: CGRect) {
+        
+        guard let context = UIGraphicsGetCurrentContext() else { return }
+        let color = TapThemeManager.colorValue(for: "CurrencyWidget.currencyDropDown.backgroundColor") ?? .white
+        
+        context.beginPath()
+        context.move(to: CGPoint(x: rect.minX, y: rect.maxY))
+        context.addLine(to: CGPoint(x: rect.minX, y: rect.minY))
+        context.addLine(to: CGPoint(x: rect.maxX, y: (rect.maxY / 2.0)))
+        context.closePath()
+        
+        context.setFillColor(color.cgColor)
+        context.fillPath()
+    }
+}
+
+
